See Description- Engine/ProtoGame2D
Refactored a lot of code
