+++ Engine/Code/Engine/Core/GameCommon.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "CoreIncludes.h"
+
+const Vec4 WINDOW_DIMS = Vec4(0.0f, 1024.0f, 0.0f, 1024.0f);
+const std::string& SHADER_XML = "../Data/Materials/shader.xml";
+const std::string& SCREENSHADER_XML = "../Data/Materials/screenShader.xml";
