+++ Engine/Code/Engine/Graphics/Buffers/FrameBuffer.cpp
@@ -34,12 +34,6 @@ void FrameBuffer::Invalidate()
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 
-	//glGenSamplers( 1, &m_ColorAttachment );
-	//glSamplerParameteri( m_ColorAttachment, GL_TEXTURE_WRAP_S, GL_REPEAT );  
-	//glSamplerParameteri( m_ColorAttachment, GL_TEXTURE_WRAP_T, GL_REPEAT );  
-	//glSamplerParameteri( m_ColorAttachment, GL_TEXTURE_MIN_FILTER, GL_LINEAR );
-	//glSamplerParameteri( m_ColorAttachment, GL_TEXTURE_MAG_FILTER, GL_LINEAR );
-
 	glBindFramebuffer(GL_FRAMEBUFFER, m_FrameBufferID);
 	glFramebufferTexture(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, m_ColorAttachment, 0);
