--- Engine/Code/Engine/Input/Mouse.cpp
@@ -1,2 +1,0 @@
-#include "Mouse.h"
-
