--- Engine/Code/Engine/Input/Mouse.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "Engine/Core/CoreIncludes.h"
-
-class Mouse
-{
-	Vec2 m_PositionThisFrame;
-	Vec2 m_PositionLastFrame;
-};
