+++ ProtoGame2D/Code/Game/Framework/App.h
@@ -1,20 +1,10 @@
 #pragma once
 
 class Game;
-class ShaderDefinition;
-class Shader;
-class FrameBuffer;
 
 class App
 {
 	bool m_IsQuitting = false;
-
-	ShaderDefinition* m_ShaderDef = nullptr;
-	ShaderDefinition* m_ScreenShaderDef = nullptr;
-	Shader* m_Shader = nullptr;
-	Shader* m_ScreenShader = nullptr;
-	FrameBuffer* m_CurrentBuffer = nullptr;
-	FrameBuffer* m_NextBuffer = nullptr;
 public:
 	Game* m_Game = nullptr;
 public:
