+++ ProtoGame2D/Code/Game/Framework/Game.h
@@ -1,7 +1,17 @@
 #pragma once
 
+class ShaderDefinition;
+class Shader;
+class FrameBuffer;
+
 class Game
 {
+	ShaderDefinition* m_ShaderDef = nullptr;
+	ShaderDefinition* m_ScreenShaderDef = nullptr;
+	Shader* m_Shader = nullptr;
+	Shader* m_ScreenShader = nullptr;
+	FrameBuffer* m_CurrentBuffer = nullptr;
+	FrameBuffer* m_NextBuffer = nullptr;
 public:
 	Game();
 	~Game();
